+++ packages/core/src/languages/makeDisposable.ts
@@ -1,8 +1,8 @@
 export function makeDisposable<T extends object>(obj: T): Disposable & T {
 	return {
-		...obj,
 		[Symbol.dispose]: () => () => {
 			// Intentionally empty to satisfy the Disposable interface.
 		},
+		...obj,
 	};
 }