+++ packages/ts/src/plugin.ts
@@ -118,6 +118,7 @@ import importEmptyBlocks from "./rules/importEmptyBlocks.ts";
 import importTypeSideEffects from "./rules/importTypeSideEffects.ts";
 import indexedObjectTypes from "./rules/indexedObjectTypes.ts";
 import instanceOfArrays from "./rules/instanceOfArrays.ts";
+import irregularWhitespaces from "./rules/irregularWhitespaces.ts";
 import isNaNComparisons from "./rules/isNaNComparisons.ts";
 import literalConstructorWrappers from "./rules/literalConstructorWrappers.ts";
 import mathMethods from "./rules/mathMethods.ts";
@@ -127,10 +128,15 @@ import moduleSpecifierLists from "./rules/moduleSpecifierLists.ts";
 import multilineAmbiguities from "./rules/multilineAmbiguities.ts";
 import namedDefaultExports from "./rules/namedDefaultExports.ts";
 import namespaceDeclarations from "./rules/namespaceDeclarations.ts";
+import namespaceImplicitAmbientImports from "./rules/namespaceImplicitAmbientImports.ts";
+import namespaceKeywords from "./rules/namespaceKeywords.ts";
 import nativeObjectExtensions from "./rules/nativeObjectExtensions.ts";
+import negativeIndexLengthMethods from "./rules/negativeIndexLengthMethods.ts";
 import negativeZeroComparisons from "./rules/negativeZeroComparisons.ts";
 import newExpressions from "./rules/newExpressions.ts";
 import newNativeNonConstructors from "./rules/newNativeNonConstructors.ts";
+import nonNullAssertedNullishCoalesces from "./rules/nonNullAssertedNullishCoalesces.ts";
+import nonNullAssertedOptionalChains from "./rules/nonNullAssertedOptionalChains.ts";
 import nonOctalDecimalEscapes from "./rules/nonOctalDecimalEscapes.ts";
 import numericLiteralParsing from "./rules/numericLiteralParsing.ts";
 import objectCalls from "./rules/objectCalls.ts";
@@ -158,6 +164,7 @@ import unsafeNegations from "./rules/unsafeNegations.ts";
 import variableDeletions from "./rules/variableDeletions.ts";
 import voidOperator from "./rules/voidOperator.ts";
 import withStatements from "./rules/withStatements.ts";
+import wrapperObjects from "./rules/wrapperObjects.ts";
 
 export const ts = createPlugin({
 	files: {
@@ -283,6 +290,7 @@ export const ts = createPlugin({
 		importTypeSideEffects,
 		indexedObjectTypes,
 		instanceOfArrays,
+		irregularWhitespaces,
 		isNaNComparisons,
 		literalConstructorWrappers,
 		mathMethods,
@@ -292,10 +300,15 @@ export const ts = createPlugin({
 		multilineAmbiguities,
 		namedDefaultExports,
 		namespaceDeclarations,
+		namespaceImplicitAmbientImports,
+		namespaceKeywords,
 		nativeObjectExtensions,
+		negativeIndexLengthMethods,
 		negativeZeroComparisons,
 		newExpressions,
 		newNativeNonConstructors,
+		nonNullAssertedNullishCoalesces,
+		nonNullAssertedOptionalChains,
 		nonOctalDecimalEscapes,
 		numericLiteralParsing,
 		objectCalls,
@@ -323,5 +336,6 @@ export const ts = createPlugin({
 		variableDeletions,
 		voidOperator,
 		withStatements,
+		wrapperObjects,
 	],
 });